QHDEV-693 - Break PrecompilePlugin.js into smaller documented functions and extract import.xml generation into its own module
